Fix sonarlint warnings

+++ jlox/src/main/java/com/charlesbourget/lox/Interpreter.java
@@ -22,7 +22,7 @@ public class Interpreter implements Expr.Visitor<Object>, Stmt.Visitor<Void> {
 
             @Override
             public Object call(Interpreter interpreter, List<Object> arguments) {
-                return (double) System.currentTimeMillis() / 1000.0;
+                return System.currentTimeMillis() / 1000.0;
             }
 
             @Override
